requirements.txt
wordcloud>=1.9.0
pkuseg>=0.0.25
Pillow>=10.0.0
numpy>=1.24.0
aiohttp>=3.9.0
